refactor: 로또 넘버 생성 전략을 LottoGenerator로 이름 변경
- 프로그램적으로 생각했을 때 전략보다는 로또번호 생성이 더 좋은 이름일거 같아서
LottoGenerator, ManualLottoGenerator, AutomaticLottoGenerator로 이름 변경
